+++ src/ts/util/selection.ts
@@ -32,7 +32,10 @@ export const getCursorPosition = (editor: HTMLElement) => {
             // markdown 模式回车
             cursorRect = children[range.startOffset].getClientRects()[0];
         } else {
-            cursorRect = (range.startContainer as HTMLElement).getClientRects()[0]; // <td></td>
+            // in table or code block
+            range.selectNode(range.startContainer.childNodes[Math.max(0, range.startOffset - 1)])
+            cursorRect = range.getClientRects()[0];
+            range.collapse(false);
         }
         if (!cursorRect) {
             let parentElement = range.startContainer.childNodes[range.startOffset] as HTMLElement;